requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
-e git+https://github.com/django-import-export/django-import-export.git@ebf58e0fe6b27285bb1c9bdfcfff0792bd7ad116#egg=django_import_export
djangorestframework==3.14.0
et-xmlfile==1.1.0
Flask==3.0.0
idna==3.6
imgkit==1.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
MouseInfo==0.1.3
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.1.3
pdfkit==1.0.0
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wikipedia==1.4.0
wkhtmltopdf==0.2
xlrd==2.0.1
xlwt==1.3.0
